put mock code in separate file to reuse elsewhere
